Trabajo_API/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
sqlmodel==0.0.21
python-multipart==0.0.9

# Security & analysis
bandit==1.7.9
semgrep==1.89.0
pip-audit==2.7.3

# Testing
httpx==0.27.0
pytest==8.2.0

argon2-cffi==23.1.0
passlib[bcrypt]==1.7.4